+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @glissade/eslint-plugin
 
+## 0.77.0-pre.0
+
 ## 0.76.0
 
 ## 0.76.0-pre.0