+++ packages/svg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/svg
 
+## 0.77.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.77.0-pre.0
+  - @glissade/core@0.77.0-pre.0
+
 ## 0.76.0
 
 ### Patch Changes